+++ man-roxygen/ex-mod_scatterplot.R
@@ -202,3 +202,48 @@ if (interactive()) {
   runApp(shinyApp(ui = ui, server = server), launch.browser = TRUE)
 }
 
+
+
+# vaccinaction index vs other variables
+if (interactive()) {
+  library(shiny)
+  library(dplyr)
+  library(tidyr)
+  library(ggplot2)
+  library(plotly)
+
+  ui <- fluidPage(
+    tagList(
+      Covid19Mirai:::golem_add_external_resources(),
+      mod_scatterplot_ui("plot", growth = FALSE, hospvars = "keep")
+    )
+  )
+  server <- function(input, output, session) {
+
+    # Data ----
+    orig_data <- get_datahub() %>%
+      get_timeseries_by_contagion_day_data()
+
+
+    pop_data = get_pop_datahub()
+    orig_data_aggregate = build_data_aggr(orig_data, pop_data)
+
+    lw_orig_data_aggregate =  lw_vars_calc(orig_data_aggregate)
+
+    orig_data_aggregate_today = orig_data_aggregate %>%
+      add_growth_death_rate()
+    orig_data_aggregate_today = orig_data_aggregate_today  %>%
+      left_join(lw_orig_data_aggregate %>% select(-population))
+
+    #inputcountries = c("Italy","Germany", "Switzerland", "Sweden", "UK", "France", "Spain", "Russia") # example with countries
+    orig_data_aggregate_today = orig_data_aggregate_today %>% filter(continent == "Europe" & population > 1000000)
+
+    inputcountries = unique(orig_data_aggregate_today$Country.Region)
+
+    callModule(mod_scatterplot_server, "plot", orig_data_aggregate_today, countries = inputcountries, nmed = 10000, wmed = 7, n_highlight = 10, istop = FALSE, growth = FALSE, fitted = FALSE, xvar = "lm_confirmed_rate_1M_pop")
+
+  }
+  runApp(shinyApp(ui = ui, server = server), launch.browser = TRUE)
+}
+
+