+++ man-roxygen/ex-mod_stackedbarplot_status.R
@@ -8,11 +8,9 @@ if (interactive()) {
     )
   )
   server <- function(input, output, session) {
-
     # Data ----
     orig_data <- get_datahub() %>%
       get_timeseries_by_contagion_day_data()
-
 
     pop_data = get_pop_datahub()
     orig_data_aggregate = build_data_aggr(orig_data, pop_data)